+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Adding `s::highlightJsLangData()`.
 - Upgrading to latest release of Ace (fork maintained by WebSharks).
 - Automatically defer JS files inserted into the footer.
+- Adding Lodash library as a loadable JS module.
 
 ## v170623.59000